+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+### 1.8.2: October 11th, 2021
 * New stat: Past Bookings by date and by type
 * New stat: Future Bookings by date
 * Fix allow admin up to 10 participants (equal to 10)